Teleporter now shows line in game
Added a line from the player to the mouse that shows in game.
This will need to be adjusted to have a max length decided later.
Need to add a faded copy of barry to the end of the line, also need to
check for collisions and adjust the ray accordingly.
Ray should eventually only show when the ability is available and only
show valid locations for barry to be placed.